Refactor: 불명확한 컴포넌트 폴더 및 파일명 수정
